Disconnect the cm_id when destroying the client instance.
We need to get the QPs flushed first by disconnecting the cm_ids and
waiting for DISCONNECTED events.  This allows the async cq completion
event thread to process the flush CQEs. Then we stop the async thread and
join it. This ensures no more polling will be done on the CQ.  Then we
destroy all the QPS and CQs via netlev_conn_free().  Finally we destroy
the completion channel.